# openapi.yaml
...
openapi: 3.1.0
info:
  title: Mapvest API
  version: 0.1.0-alpha.0
  description: Mapvest turns places and objects into investable tickers. All request and response shapes are generated from zod schemas in `packages/core/src/schemas` — never hand-edit `openapi.yaml`.
  contact:
    name: Mapvest
    url: https://github.com/jawauntb/mapvest
  license:
    name: UNLICENSED
servers:
  - url: http://localhost:3001
    description: Local dev
  - url: https://mapvest-api.up.railway.app
    description: Railway (staging/production)
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: "Session token issued by `POST /v1/auth/session`. Attach as `Authorization: Bearer <token>`."
  schemas:
...
    QuotaExceeded:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
          enum:
            - quota_exceeded
        remaining:
          type: integer
          minimum: 0
        limit:
          type: integer
          minimum: 0
        priceUsd:
          type: number
        interval:
          type: string
          enum:
            - month
      required:
        - error
        - code
        - remaining
        - limit
...
    CompanyGraphResponse:
      type: object
      properties:
        ticker:
          type: string
        edges:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              srcTicker:
                type: string
              dstTicker:
                type: string
              dstName:
                type: string
              edgeType:
                type: string
                enum:
                  - supplies
                  - buys_from
                  - competes_with
                  - complements
              weight:
                type: number
                minimum: 0
                maximum: 1
              reasoning:
                type: string
              sources:
                type: array
                items:
                  type: object
                  properties:
                    provider:
                      type: string
                      enum:
                        - exa
                        - openrouter
                        - gemini
                        - massive
                        - yahoo
                        - polygon
                        - sec
                        - fred
                        - manual
                    url:
                      type: string
                      format: uri
                    fetchedAt:
                      type: string
                    confidence:
                      type: string
                      enum:
                        - high
                        - medium
                        - low
                  required:
                    - provider
                    - fetchedAt
                    - confidence
              asOf:
                type: string
              createdAt:
                type: string
            required:
              - id
              - srcTicker
              - dstName
              - edgeType
              - weight
              - reasoning
              - sources
              - createdAt
        count:
          type: number
        generatedAt:
          type: string
        sources:
          type: array
          items:
            type: object
            properties:
              provider:
                type: string
                enum:
                  - exa
                  - openrouter
                  - gemini
                  - massive
                  - yahoo
                  - polygon
                  - sec
                  - fred
                  - manual
              url:
                type: string
                format: uri
              fetchedAt:
                type: string
              confidence:
                type: string
                enum:
                  - high
                  - medium
                  - low
            required:
              - provider
              - fetchedAt
              - confidence
      required:
        - ticker
        - edges
        - count
        - generatedAt
        - sources
...
    EnvironmentBrief:
      type: object
      properties:
        sector:
          type: string
        headline:
          type: string
        body:
          type: string
        tailwinds:
          type: array
          items:
            type: string
        headwinds:
          type: array
          items:
            type: string
        series:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              label:
                type: string
              latest:
                type: number
              unit:
                type: string
              asOf:
                type: string
            required:
              - id
              - label
              - latest
              - asOf
        generatedAt:
          type: string
        sources:
          type: array
          items:
            type: object
            properties:
              provider:
                type: string
                enum:
                  - exa
                  - openrouter
                  - gemini
                  - massive
                  - yahoo
                  - polygon
                  - sec
                  - fred
                  - manual
              url:
                type: string
                format: uri
              fetchedAt:
                type: string
              confidence:
                type: string
                enum:
                  - high
                  - medium
                  - low
            required:
              - provider
              - fetchedAt
              - confidence
      required:
        - sector
        - headline
        - body
        - tailwinds
        - headwinds
        - series
        - generatedAt
        - sources
...
paths:
...
  /v1/graph/{ticker}:
    get:
      summary: Company value-chain graph
      description: "Cache-first: served from the `company_edges` store, extracted on miss with in-flight dedupe per ticker and refreshed when a new 10-K lands rather than on a short TTL. Cache hits are free and identity-less; a cache MISS spends provider money, so generation requires a bearer session or `X-Device-Id` and counts against the free-tier generation quota (402 when spent; concurrent callers sharing one in-flight extraction are charged once). Edges are cited from 10-K items 1/1A evidence (supplier concentration and >10% customers are disclosed there) plus Exa open-web results; every edge carries `sources` (AGENTS.md §6) and an edge that cannot be cited is not emitted. Private counterparties keep `dstName` with no `dstTicker` — never an invented one. Empty and failing extractions are negatively cached for an hour. Returns 502 when extraction fails."
      tags:
        - finance
      parameters:
        - schema:
            type: string
            example: NVDA
          required: true
          name: ticker
          in: path
      responses:
        "200":
          description: Graph edges returned (cache hit or fresh extraction).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyGraphResponse"
        "400":
          description: Bad request — malformed ticker, or a cache-miss generation attempted with neither a bearer session nor an `X-Device-Id` header.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                required:
                  - error
        "402":
          description: Free-tier generation quota spent. Clients must present the paywall, not a generic error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuotaExceeded"
        "429":
          description: Rate limit exceeded.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                required:
                  - error
        "502":
          description: Graph extraction failed (filings or Exa evidence unavailable).
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                required:
                  - error
...
  /v1/environment/{sector}:
    get:
      summary: Macro environment brief for a sector
      description: "The gather → LLM → tailwinds/headwinds brief at sector scale, cached 24h. `series` holds the quantitative FRED observations behind the brief; policy and culture color gathered via Exa with recency filters is qualitative and is cited in `sources` alongside them, never promoted into `series`. Metered exactly like `/v1/graph` — cache hits are free, a cache MISS requires a bearer session or `X-Device-Id` and counts against the free-tier generation quota (402 when spent). Degrades honestly: with `FRED_API_KEY` unset the brief ships with `series: []` and Exa color only; 503 only when neither FRED nor Exa (nor the LLM) is configured. Macro numbers are never synthesized."
      tags:
        - finance
      parameters:
        - schema:
            type: string
            example: Consumer Staples
          required: true
          name: sector
          in: path
      responses:
        "200":
          description: Environment brief returned (cache hit or fresh generation).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvironmentBrief"
        "400":
          description: Bad request — unknown sector, or a cache-miss generation attempted with neither a bearer session nor an `X-Device-Id` header.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                required:
                  - error
        "402":
          description: Free-tier generation quota spent. Clients must present the paywall, not a generic error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuotaExceeded"
        "429":
          description: Rate limit exceeded.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                required:
                  - error
        "502":
          description: Environment brief generation failed.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                required:
                  - error
        "503":
          description: Environment layer not configured (neither FRED_API_KEY nor EXA_API_KEY, or no LLM key).
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                required:
                  - error
...